ejercicio1 y 2

+++ src/tema2/ejercicio2.java
@@ -6,9 +6,9 @@ public class ejercicio2 {
     public static void main(String[] args) {
         System.out.println("introduce la base del triangulo");
         Scanner in = new Scanner(System.in);
-        int base = in.nextInt();
+        double base = in.nextDouble();
         System.out.println("introduce la altura del traingulo");
-        int altura= in.nextInt();
+        double altura= in.nextDouble();
         double superficie= (base*altura)/2;
         System.out.println("la superficie del triangulo es " +superficie);
     }
